Higehr randonmess
OKAY SO
It was generating numbers between 1 and 999, and for my test folder of 481 images several numbers were popping up more than once so they'd get deleted. I just upped the randomness to between 1 and 999,999 because that's probably fine. If you're doing more like 25% of 999,999 in one go just up the range more. I'll fix it properly later.